backend/pyproject.toml
[project]
name = "jobsletter"
version = "0.1.0"
description = "A tool to scrape career pages and send yourself alerts and a newsletter."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "bcrypt>=4.3.0",
    "beautifulsoup4>=4.13.3",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "fastapi-cli>=0.0.7",
    "pg8000>=1.31.2",
    "pydantic-extra-types>=2.10.3",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.34.0",
]
